Cleanup: move Enum dropdowsn from form/common to common/Dropdowns
