+++ ui/src/components/forms/route/DirectionDropdown.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next';
 import { mapDirectionToUiName } from '../../../i18n/uiNameMappings';
 import { RouteDirection } from '../../../types/RouteDirection';
 import { FormInputProps } from '../../common';
-import { EnumDropdown } from '../common/EnumDropdown';
+import { EnumDropdown } from '../../common/Dropdowns';
 
 type DirectionDropdownProps = FormInputProps & {
   readonly testId?: string;
