+++ ui/src/components/forms/timetables/LineTypeMultiSelectDropdown.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next';
 import { RouteTypeOfLineEnum } from '../../../generated/graphql';
 import { mapLineTypeToUiName } from '../../../i18n/uiNameMappings';
 import { FormInputProps } from '../../common';
-import { EnumMultiSelectDropdown } from '../common/EnumMultiSelectDropdown';
+import { EnumMultiSelectDropdown } from '../../common/Dropdowns';
 
 type LineTypeMultiSelectDropdownProps = FormInputProps & {
   readonly id?: string;
