+++ ui/src/components/stop-registry/components/TerminalTypeDropdown/TerminalTypeDropdown.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react';
 import { useTranslation } from 'react-i18next';
 import { mapTerminalTypeToUiName } from '../../../../i18n/uiNameMappings';
 import { FormInputProps } from '../../../common';
-import { EnumDropdown } from '../../../forms/common';
+import { EnumDropdown } from '../../../common/Dropdowns';
 import { TerminalType } from '../../types/TerminalType';
 
 type TerminalTypeDropdownProps = FormInputProps;
