+++ ui/src/components/stop-registry/search/components/StopSearchBar/BasicFilters/SearchForDropdown.tsx
@@ -2,8 +2,9 @@ import pick from 'lodash/pick';
 import { FC } from 'react';
 import { useController, useFormContext } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
+import { EnumDropdown } from '../../../../../common/Dropdowns';
 import { Column } from '../../../../../common/LayoutComponents';
-import { EnumDropdown, InputLabel } from '../../../../../forms/common';
+import { InputLabel } from '../../../../../forms/common';
 import { SearchFor, StopSearchFilters, defaultFilters } from '../../../types';
 import { trSearchFor } from '../../../utils';
 import { stopSearchBarTestIds } from '../stopSearchBarTestIds';
