+++ ui/src/components/stop-registry/stops/stop-details/basic-details/basic-details-form/StopOtherDetailsFormRow.tsx
@@ -4,8 +4,9 @@ import { useTranslation } from 'react-i18next';
 import { mapStopRegistryTransportModeTypeToUiName } from '../../../../../../i18n/uiNameMappings';
 import { JoreStopRegistryTransportModeType } from '../../../../../../types/stop-registry';
 import { SimpleButton } from '../../../../../common/Buttons';
+import { EnumDropdown } from '../../../../../common/Dropdowns';
 import { Column, Row } from '../../../../../common/LayoutComponents';
-import { EnumDropdown, FormRow, InputField } from '../../../../../forms/common';
+import { FormRow, InputField } from '../../../../../forms/common';
 import { ChooseTimingPlaceDropdown } from '../../../../../forms/stop/ChooseTimingPlaceDropdown';
 import { StopBasicDetailsFormState } from './schema';
 
@@ -92,3 +93,4 @@ export const StopOtherDetailsFormRow: FC<StopOtherDetailsFormRowProps> = ({
     </FormRow>
   );
 };
+export default StopOtherDetailsFormRow;
