+++ ui/src/components/stop-registry/terminals/components/info-spots/terminal-info-spots-form/TerminalInfoSpotsFormFields.tsx
@@ -11,9 +11,9 @@ import {
 } from '../../../../../../i18n/uiNameMappings';
 import { EnrichedParentStopPlace } from '../../../../../../types';
 import { AddNewButton, SimpleButton } from '../../../../../common/Buttons';
+import { EnumDropdown } from '../../../../../common/Dropdowns';
 import { Column, Row } from '../../../../../common/LayoutComponents';
 import {
-  EnumDropdown,
   InputField,
   NullableBooleanDropdown,
 } from '../../../../../forms/common';
